[SPARK-37156][PYTHON] Inline type hints for python/pyspark/storagelevel.py
### What changes were proposed in this pull request?
Inline type hints for python/pyspark/storagelevel.py

### Why are the changes needed?
We can take advantage of static type checking within the functions by inlining the type hints.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests

Closes #34437 from dchvn/SPARK-37156.

Authored-by: dchvn <dgd_contributor@viettel.com.vn>
